fixing: username removed from webRtcClient
