+++ feature/calls/src/main/java/com/zhigaras/calls/domain/CallsController.kt
@@ -1,11 +1,10 @@
 package com.zhigaras.calls.domain
 
-import com.zhigaras.calls.domain.model.ConnectionData
 import org.webrtc.SurfaceViewRenderer
 
 interface CallsController {
     
-    fun startNegotiation(opponentId: String)
+    fun startNegotiation(opponentId: String, userId: String)
     
     fun subscribeToConnectionEvents(userId: String)
